Refractor Validation.py
Move following functions:
validate_csv_extension
validate_output_directory
validate_filename_uniqueness
handle_missing_filenames
setup_expected_columns
